Inline jsoncharutils per-implementation
