+++ src/haswell/dom_parser_implementation.cpp
@@ -1,5 +1,6 @@
 #include "haswell/begin_implementation.h"
 #include "haswell/dom_parser_implementation.h"
+#include "generic/stage2/jsoncharutils.h"
 
 //
 // Stage 1
